+++ cloud_proyecto1/src/main/java/microservicios/cloud_proyecto1/categoria/infrastructure/CategoriaRepository.java
@@ -3,5 +3,6 @@ package microservicios.cloud_proyecto1.categoria.infrastructure;
 import microservicios.cloud_proyecto1.categoria.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+
 }
